a shitload of Encode/Decode/TypeInfo
feat: add SCALE Codec to blockifier repo

Refers: https://github.com/keep-starknet-strange/madara/issues/1430

feat: add SCALE Codec to Sierra Types

Refers: https://github.com/keep-starknet-strange/madara/issues/1430

wip: move SCALE Codec changes from Madara repository

Refers: https://github.com/keep-starknet-strange/madara/issues/1430

feat: add SCALE Codec to transaction types

Refers https://github.com/keep-starknet-strange/madara/issues/1430

chore: update SCALE Codec dependencies

Refers: https://github.com/keep-starknet-strange/madara/issues/1430

feat: add SCALE Codec to CommitmentStateDiff

Refers: https://github.com/keep-starknet-strange/madara/issues/1430

feat: scale-info

feat: impl Default for ContractClassV1

chore: encode/decode/scale-info for BlockInfo & GasPrices

chore: add de/serde for GasPrice

add more decode + scale for versioned_consts

+++ crates/blockifier/src/execution/deprecated_entry_point_execution.rs
@@ -126,7 +126,7 @@ pub fn resolve_entry_point_pc(
                 if entry_point.selector
                     == EntryPointSelector(StarkHash::from(DEFAULT_ENTRY_POINT_SELECTOR))
                 {
-                    return Ok(entry_point.offset.0);
+                    return Ok(entry_point.offset.0 as usize);
                 } else {
                     return Err(PreExecutionError::EntryPointNotFound(call.entry_point_selector));
                 }
@@ -148,7 +148,7 @@ pub fn resolve_entry_point_pc(
     let entry_point = filtered_entry_points
         .first()
         .expect("The number of entry points with the given selector is exactly one.");
-    Ok(entry_point.offset.0)
+    Ok(entry_point.offset.0 as usize)
 }
 
 pub fn prepare_call_arguments(
