+++ crates/blockifier/src/fee/actual_cost.rs
@@ -179,7 +179,7 @@ impl<'a> ActualCostBuilder<'a> {
 
         // Add reverted steps to actual_resources' n_steps for correct fee charge.
         *actual_resources.0.get_mut(&abi_constants::N_STEPS_RESOURCE.to_string()).unwrap() +=
-            self.n_reverted_steps;
+            self.n_reverted_steps as u128;
 
         let tx_info = &self.tx_context.tx_info;
         let actual_fee = if tx_info.enforce_fee()?
